chore: organize mocks
