+++ src/workers/webdav/modules/files/test/__mocks__/environment/UploadStrategyFunciontMock.ts
